Setup folder structure for Webpack
