--- frontend/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link rel="stylesheet" href="styles/style.css" />
-    <title>RPG Attack Roller</title>
-  </head>
-  <body>
-    <div class="flex-wrap">
-      <div class="app-card">
-        <h1 id="playerName">Unknown</h1>
-        <h2> Racial STR: <span id="racialStr"></span></h2>
-        <h2>Racial DEX: <span id="racialDex"></span></h2>
-        <h2>Total bonus Str: <span id="bonusStr"></span></h2>
-        <div class="race-button-group" id="raceSelections"></div>
-
-        <div class="optgroups">
-          <label for='weapon-select'>Weapon Selection</label>
-          <select name="weapon-select" id="weapon-select"></select>
-        </div>
-        <label> Enter your Strength: </label>
-        <input type="text" class="input-strength">
-        <button id="attackRollBtn" disabled>Roll Attack</button>
-        <h3 id="rollResult">Roll: </h3>
-        <h3 id="totalResult">Totalroll: </h3>
-      </div>
-    </div>
-   
-
-    <script src="app.js"></script>
-    <script src="weapon.js"></script>
-  </body>
-</html>
